L30 burn added + Pigeons now appear on dashboard

+++ src/main/java/com/sandy/sconsole/core/util/DayValue.java
@@ -2,4 +2,4 @@ package com.sandy.sconsole.core.util;
 
 import java.util.Date;
 
-public record DayValue( Date date, float value ) {}
+public record DayValue( Date date, double value ) {}
